requirements.txt
#psycopg2
flask
psycopg2-binary
azure-identity
azure-keyvault-secrets
gunicorn
azure-appconfiguration
azure-appconfiguration-provider
requests